hola/requirements.txt
flask
openai==0.28.0
python-docx
pillow
PyPDF2
dotenv